Changed audio to video in The Video Tag section
I have changed audio to video at multiple places.  Also changed AT to TO in "skip TO a specific position.  This is my first attempt at forking and making changes and am not sure of the exact procedure and have used a YouTube video as reference; hence, please excuse if I goof it up.